Move testutils.ts to helpers.ts
