+++ package.json
@@ -8,6 +8,7 @@
   "browserslist": [
     "maintained node versions"
   ],
+  "main": "build/index.js",
   "scripts": {
     "build": "tsc -p tsconfig.json",
     "fix": "run-s fix:*",
