+++ src/lib/ibcclient.spec.ts
@@ -2,8 +2,6 @@ import { sleep } from '@cosmjs/utils';
 import test from 'ava';
 import { MsgTransfer } from 'cosmjs-types/ibc/applications/transfer/v1/tx';
 
-import { buildCreateClientArgs, prepareConnectionHandshake } from './ibcclient';
-import { Link } from './link';
 import {
   gaia,
   ics20,
@@ -11,7 +9,9 @@ import {
   setup,
   TestLogger,
   wasmd,
-} from './testutils';
+} from './helpers';
+import { buildCreateClientArgs, prepareConnectionHandshake } from './ibcclient';
+import { Link } from './link';
 import {
   buildClientState,
   buildConsensusState,
