+++ src/lib/index.ts
@@ -2,5 +2,5 @@ export { AckWithMetadata, Endpoint } from './endpoint';
 export { IbcClient } from './ibcclient';
 export { Link, RelayInfo, RelayedHeights } from './link';
 export { Logger, NoopLogger } from './logger';
-export * as testutils from './testutils';
-export { CosmWasmSigner } from './testutils';
+export * as testutils from './helpers';
+export { CosmWasmSigner } from './helpers';
