+++ src/lib/osmosis.spec.ts
@@ -7,7 +7,7 @@ import {
   osmosis,
   signingClient,
   TestLogger,
-} from './testutils';
+} from './helpers';
 
 test.serial('funds account and checks balance', async (t) => {
   const logger = new TestLogger();
